requirements.txt
django
djangorestframework
django-password-validators